requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.2
mdurl==0.1.2
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
pyyaml==6.0.2
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.6.4
sentry-sdk==2.38.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.48.0
typer==0.19.1
typing-extensions==4.15.0
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.36.0
watchfiles==1.1.0
websockets==15.0.1
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.0.1
sqlalchemy==2.0.27
